+ Update core modules.
 - Add missing LineDash implementation.
 - Fix slow ticker.
 - Update ellipse drawing logic.
 - Abstract drawing logical to independent module.
