--- packages/core/src/core/models/svg.ts
@@ -1,8 +1,0 @@
-interface ISvgCommand {
-  marker: string
-  values: number[]
-}
-
-export {
-  ISvgCommand
-}
